split routes into its their file
